refactor: isolate a snarkos-node-network crate
