+++ node/network/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "snarkos-node-router-messages"
+name = "snarkos-node-network"
 version = "4.2.2"
 authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "Node messages for a decentralized operating system"
+description = "Network-related objects and traits for nodes"
 homepage = "https://aleo.org"
 repository = "https://github.com/ProvableHQ/snarkOS"
 keywords = [
@@ -17,41 +17,38 @@ license = "Apache-2.0"
 edition = "2024"
 
 [features]
-default = [ ]
 test = [ ]
 
-[dependencies.bytes]
+[dependencies.aleo-std]
 workspace = true
 
-[dependencies.snarkos-node-network]
+[dependencies.anyhow]
 workspace = true
 
-[dependencies.snarkos-node-bft-events]
+[dependencies.locktick]
 workspace = true
+features = [ "parking_lot" ]
+optional = true
 
-[dependencies.snarkos-node-sync-locators]
+[dependencies.parking_lot]
 workspace = true
-
-[dependencies.snarkvm]
-workspace = true
-
-[dependencies.tokio-util]
-workspace = true
-features = [ "codec" ]
 
 [dependencies.tracing]
 workspace = true
 
-[dev-dependencies.snarkos-node-sync-locators]
+[dependencies.serde]
 workspace = true
+
+[dependencies.snarkos-node-tcp]
+workspace = true
+
+[dependencies.snarkvm]
+workspace = true
+features = [ "console" ]
+
+[dependencies.tokio]
+workspace = true
+
+[dev-dependencies.snarkos-node-network]
+path = "."
 features = [ "test" ]
-
-[dev-dependencies.snarkvm]
-workspace = true
-features = [ "algorithms", "test-helpers" ]
-
-[dev-dependencies.proptest]
-workspace = true
-
-[dev-dependencies.test-strategy]
-workspace = true
