+++ node/router/tests/banning.rs
@@ -16,7 +16,7 @@
 mod common;
 use common::*;
 
-use snarkos_node_router::PeerPoolHandling;
+use snarkos_node_network::PeerPoolHandling;
 use snarkos_node_tcp::{
     P2P,
     protocols::{Disconnect, Handshake},
