+++ node/src/bootstrap_client/network.rs
@@ -20,13 +20,9 @@ use crate::{
         events::{self, Event},
     },
     bootstrap_client::codec::BootstrapClientCodec,
+    network::{ConnectionMode, NodeType, Peer, PeerPoolHandling, Resolver},
     router::{
-        ConnectionMode,
         MAX_PEERS_TO_SEND,
-        NodeType,
-        Peer,
-        PeerPoolHandling,
-        Resolver,
         messages::{self, Message},
     },
     tcp::{ConnectionSide, P2P, Tcp, protocols::*},
