+++ node/src/traits.rs
@@ -13,7 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use snarkos_node_router::{PeerPoolHandling, Routing, messages::NodeType};
+use snarkos_node_network::{NodeType, PeerPoolHandling};
+use snarkos_node_router::Routing;
 use snarkvm::prelude::{Address, Network, PrivateKey, ViewKey};
 
 use once_cell::sync::OnceCell;
